backend/requirements.txt
python-dotenv==1.0.1
tinydb==4.8.0
python-multipart==0.0.9
httpx==0.27.0
beautifulsoup4==4.12.3
openai==1.52.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
slowapi==0.1.9
anthropic
pytest
pytest-asyncio
pytest-httpx
